pyproject.toml
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.

[tool.poetry]
name = "otava"
version = "0.6.0"
description = ""
authors = ["Piotr Kołaczkowski <pkolaczk@datastax.com>"]

[tool.poetry.dependencies]
dateparser = "^1.0.0"
expandvars = "^0.6.5"
numpy = "1.24"
python = ">=3.8,<3.13"
python-dateutil = "^2.8.1"
signal-processing-algorithms = "^1.3.2"
"ruamel.yaml" = "=0.17.21"
requests = "^2.25.1"
pystache = "^0.6.0"
tabulate = "^0.8.7"
validators = "^0.18.2"
slack-sdk = "^3.4.2"
google-cloud-bigquery = "^3.25.0"
pg8000 = "^1.31.2"


[tool.poetry.group.dev.dependencies]
pytest = "^6.2.2"
pytest-benchmark = "^4.0.0"
pytz = "2021.1"
tox = "^3.25.0"
flake8 = "^4.0.1"
autoflake = "^1.4"
isort = "^5.10.1"
ruff = "^0.6.9"
pre-commit = "3.5.0"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:future",
    "ignore::pytest.PytestCollectionWarning"
]

[tool.poetry.scripts]
otava = 'otava.main:main'

[tool.ruff]
line-length = 100

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
